Add new interaction data from STRING-DB to b sub
Fixed bug in external links for B. subtilis and added new interaction data. It seems that we are missing annotations for the new interactions because I tried to run the import with our current GO terms, but no records were changed. I will see if I can get new annotation data for the new interactions we imported.